+++ src/Transformer/MetadataUtil/ClassMetadataFactory/ClassMetadataFactory.php
@@ -11,10 +11,14 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-namespace Rekalogika\Mapper\Transformer\MetadataUtil;
+namespace Rekalogika\Mapper\Transformer\MetadataUtil\ClassMetadataFactory;
 
 use Rekalogika\Mapper\Transformer\EagerPropertiesResolver\EagerPropertiesResolverInterface;
+use Rekalogika\Mapper\Transformer\MetadataUtil\AttributesExtractorInterface;
+use Rekalogika\Mapper\Transformer\MetadataUtil\ClassMetadataFactoryInterface;
+use Rekalogika\Mapper\Transformer\MetadataUtil\DynamicPropertiesDeterminerInterface;
 use Rekalogika\Mapper\Transformer\MetadataUtil\Model\ClassMetadata;
+use Rekalogika\Mapper\Transformer\MetadataUtil\UnalterableDeterminerInterface;
 use Rekalogika\Mapper\Util\ClassUtil;
 
 /**
@@ -24,9 +28,9 @@ final readonly class ClassMetadataFactory implements ClassMetadataFactoryInterfa
 {
     public function __construct(
         private EagerPropertiesResolverInterface $eagerPropertiesResolver,
-        private DynamicPropertiesDeterminer $dynamicPropertiesDeterminer,
-        private AttributesExtractor $attributesExtractor,
-        private UnalterableDeterminer $unalterableDeterminer,
+        private DynamicPropertiesDeterminerInterface $dynamicPropertiesDeterminer,
+        private AttributesExtractorInterface $attributesExtractor,
+        private UnalterableDeterminerInterface $unalterableDeterminer,
     ) {}
 
     /**