+++ src/Transformer/MetadataUtil/PropertyAccessInfoExtractor/PropertyAccessInfoExtractor.php
@@ -11,8 +11,9 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-namespace Rekalogika\Mapper\Transformer\MetadataUtil;
+namespace Rekalogika\Mapper\Transformer\MetadataUtil\PropertyAccessInfoExtractor;
 
+use Rekalogika\Mapper\Transformer\MetadataUtil\PropertyAccessInfoExtractorInterface;
 use Symfony\Component\PropertyInfo\PropertyReadInfo;
 use Symfony\Component\PropertyInfo\PropertyReadInfoExtractorInterface;
 use Symfony\Component\PropertyInfo\PropertyWriteInfo;
@@ -21,88 +22,38 @@ use Symfony\Component\PropertyInfo\PropertyWriteInfoExtractorInterface;
 /**
  * @internal
  */
-final class PropertyAccessInfoExtractor
+final readonly class PropertyAccessInfoExtractor implements PropertyAccessInfoExtractorInterface
 {
-    /**
-     * @var array<class-string,array<string,PropertyReadInfo|false>>
-     */
-    private array $readInfoCache = [];
-
-    /**
-     * @var array<class-string,array<string,PropertyWriteInfo|false>>
-     */
-    private array $writeInfoCache = [];
-
-    /**
-     * @var array<class-string,array<string,PropertyWriteInfo|false>>
-     */
-    private array $constructorInfoCache = [];
-
     public function __construct(
         private PropertyReadInfoExtractorInterface $propertyReadInfoExtractor,
         private PropertyWriteInfoExtractorInterface $propertyWriteInfoExtractor,
     ) {}
 
-    /**
-     * @param class-string $class
-     */
     public function getReadInfo(
         string $class,
         string $property,
     ): ?PropertyReadInfo {
-        $readInfo = $this->readInfoCache[$class][$property] ?? null;
-
-        if ($readInfo !== null) {
-            return $readInfo === false ? null : $readInfo;
-        }
-
-        $readInfo = $this->propertyReadInfoExtractor
+        return $this->propertyReadInfoExtractor
             ->getReadInfo($class, $property);
-
-        $this->readInfoCache[$class][$property] = $readInfo ?? false;
-
-        return $readInfo;
     }
 
-    /**
-     * @param class-string $class
-     */
     public function getWriteInfo(
         string $class,
         string $property,
     ): ?PropertyWriteInfo {
-        $writeInfo = $this->writeInfoCache[$class][$property] ?? null;
-
-        if ($writeInfo !== null) {
-            return $writeInfo === false ? null : $writeInfo;
-        }
-
-        $writeInfo = $this->propertyWriteInfoExtractor->getWriteInfo(
+        return $this->propertyWriteInfoExtractor->getWriteInfo(
             class: $class,
             property: $property,
             context: [
                 'enable_constructor_extraction' => false,
             ],
         );
-
-        $this->writeInfoCache[$class][$property] = $writeInfo ?? false;
-
-        return $writeInfo;
     }
 
-    /**
-     * @param class-string $class
-     */
     public function getConstructorInfo(
         string $class,
         string $property,
     ): ?PropertyWriteInfo {
-        $constructorInfo = $this->constructorInfoCache[$class][$property] ?? null;
-
-        if ($constructorInfo !== null) {
-            return $constructorInfo === false ? null : $constructorInfo;
-        }
-
         $constructorInfo = $this->propertyWriteInfoExtractor->getWriteInfo(
             class: $class,
             property: $property,
@@ -118,12 +69,8 @@ final class PropertyAccessInfoExtractor
             $constructorInfo === null
             || $constructorInfo->getType() !== PropertyWriteInfo::TYPE_CONSTRUCTOR
         ) {
-            $this->constructorInfoCache[$class][$property] = false;
-
             return null;
         }
-
-        $this->constructorInfoCache[$class][$property] = $constructorInfo;
 
         return $constructorInfo;
     }